style: forgo unsigned types. closes #22

+++ capsulerun/include/capsulerun_types.h
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <stdint.h>
+
 typedef struct video_format_s {
   int width;
   int height;
   capsule_pix_fmt_t format;
   int vflip;
-  int pitch;
+  intptr_t pitch;
 } video_format_t;
 
 typedef struct audio_format_s {
